+++ tests/integration_test.rs
@@ -28,7 +28,7 @@ fn test_basic_module_instantiation() {
 
     // TCP Proxy
     let tcp_addr = SocketAddr::from(([127, 0, 0, 1], 8443));
-    let _tcp_proxy = TcpProxy::new(tcp_addr, std::collections::HashMap::new());
+    let _tcp_proxy = TcpProxy::new(tcp_addr, s3_proxy::connection_pool::EndpointOverrides::from_config(&std::collections::HashMap::new()));
 
     // Cache Manager
     let cache_dir = PathBuf::from("/tmp/test_cache");
@@ -89,10 +89,10 @@ fn test_compression_threshold() {
 fn test_tcp_proxy_creation() {
     // Test that TCP proxy can be created with different addresses
     let tcp_addr1 = SocketAddr::from(([127, 0, 0, 1], 443));
-    let _tcp_proxy1 = TcpProxy::new(tcp_addr1, std::collections::HashMap::new());
+    let _tcp_proxy1 = TcpProxy::new(tcp_addr1, s3_proxy::connection_pool::EndpointOverrides::from_config(&std::collections::HashMap::new()));
 
     let tcp_addr2 = SocketAddr::from(([0, 0, 0, 0], 8443));
-    let _tcp_proxy2 = TcpProxy::new(tcp_addr2, std::collections::HashMap::new());
+    let _tcp_proxy2 = TcpProxy::new(tcp_addr2, s3_proxy::connection_pool::EndpointOverrides::from_config(&std::collections::HashMap::new()));
 
     // Verify that the proxy can be created without errors
     // The actual functionality will be tested in integration tests